# openapi/portfolio.yaml
openapi: 3.1.0
info:


  title: Adithya Krishnan Portfolio API
  description: >
    Full API surface for the portfolio site: portfolio data
    (GitHub stats, Notion-backed blogs/projects, Spotify now-playing,
    aggregated stats), the expense tracker (Notion-backed), and service
    health/monitoring. All protected routes use a single static Bearer API key
    via the `Authorization: Bearer <API_KEY>` header.
  version: 6.0.0
servers:
  - url: https://www.adithyakrishnan.com
security:
  - bearerAuth: []
...
paths:
...
  /api/spotify:
    get:
      operationId: getNowPlaying
      summary: Currently/recently playing track
      description: >
        Returns the currently playing track, or the most recently played
        track with `isPlaying: false` if nothing is playing. Auth is
        optional here — an `Authorization` header is only validated if present.
      tags: [Portfolio Data]
      security:
        - {}
        - bearerAuth: []
      responses:
        "200":
          description: Now-playing data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NowPlaying"
        "401":
          description: An Authorization header was sent but was invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/stats:
    get:
      operationId: getAggregatedStats
      summary: Aggregated dashboard stats
      description: >
        Spotify now-playing, GitHub summary, and Notion content counts,
        fetched concurrently via `Promise.allSettled` — one failing source
        degrades gracefully instead of failing the whole response.
      tags: [Portfolio Data]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Aggregated stats.
          content:
            application/json:
              schema:
                type: object
                properties:
                  spotify:
                    $ref: "#/components/schemas/NowPlaying"
                  github:
                    type: [object, "null"]

                    properties:
                      username:
                        type: string
                      publicRepos:
                        type: integer
                      followers:
                        type: integer
                      totalStars:
                        type: integer
                  notion:
                    type: object
                    properties:
                      blogsCount:
                        type: integer
                      projectsCount:
                        type: integer
                      experiencesCount:
                        type: integer
                  availability:
                    $ref: "#/components/schemas/AvailabilityStatus"
        "401":
          $ref: "#/components/responses/OAuthUnauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/calendar:
    get:
      operationId: getCalendarEvents
      summary: Google Calendar upcoming events
      description: Returns a list of Google Calendar events for the next 7 days, along with current availability status.
      tags: [Portfolio Data]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Calendar events and status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  availability:
                    $ref: "#/components/schemas/AvailabilityStatus"
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/CalendarEvent"
        "401":
          $ref: "#/components/responses/OAuthUnauthorized"
        "500":
          $ref: "#/components/responses/ServerError"


  # ---------------------------------------------------------------------
  # Expenses — reads
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Static API key used for all protected endpoints.
        Send as `Authorization: Bearer <API_KEY>`.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "`title` and `amount` are required."
    AvailabilityStatus:
      type: object
      properties:
        status:
          type: string
          enum: [Available, Busy]
          example: "Available"
        currentEvent:
          type: [string, "null"]
          example: "Meeting with Client"
    CalendarEvent:
      type: object
      properties:
        summary:
          type: string
          example: "Sync Meeting"
        start:
          type: string
          format: date-time
          example: "2026-07-15T10:00:00Z"
        end:
          type: string
          format: date-time
          example: "2026-07-15T11:00:00Z"
        isBusy:
          type: boolean
          example: true

    # --- Portfolio data ---
...
    NowPlaying:
      type: object
      properties:
        isPlaying:
          type: boolean
        title:
          type: string
        artist:
          type: string
        songUrl:
          type: string
          format: uri
        albumImageUrl:
          type: string
          format: uri
        lastPlayedAt:
          type: [string, "null"]
          format: date-time

    # --- Expenses ---
...
  responses:
...
    OAuthUnauthorized:
      description: Missing or invalid OAuth bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Unauthorized
            message: "Valid OAuth Bearer token required"
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"